Moves config to a different dir from public
